pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "yamlsan"
version = "0.1.0"
description = "YAML Frontmatter Schema Validator -- validate markdown frontmatter against user-defined schemas"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "fuku", email = "noreply@fuku.dev" },
]
keywords = ["yaml", "frontmatter", "schema", "validator", "markdown", "cli", "obsidian"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Quality Assurance",
    "Topic :: Text Processing :: Markup :: Markdown",
]
dependencies = [
    "pyyaml>=6.0",
    "rich>=13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]

[project.scripts]
yamlsan = "yamlsan.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.coverage.run]
source = ["yamlsan"]
omit = ["tests/*"]
